separating all tests into separate files based on components
