+++ frontend/src/__tests__/Teammember.test.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import Teammember from '../components/Team-member';
 import { exportAllDeclaration } from '@babel/types';
 import {getFirstName, validateRole, trimBio, checkAuthority, generateUserID, filterBadWords, generateInfoArray, randomNumber, makeRolesBinary} from '../components/Team-member'; 
+import TestRenderer from 'react-test-renderer';
 
 const testString = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla congue sodales semper. Quisque posuere justo eu semper lacinia. Mauris at orci posuere, mollis erat nec, luctus dolor. Donec consequat posuere arcu id interdum. Praesent eleifend malesuada elit, et varius odio vulputate sit amet. Suspendisse potenti. Nulla turpis ex, sollicitudin a libero et, ullamcorper sollicitudin mi. Nulla efficitur dui neque, at tincidunt metus tempor consequat. Aenean id tempor orci. Ut aliquet cursus urna, sed malesuada sem suscipit in. In aliquam ullamcorper purus, quis sagittis purus viverra id. Mauris eget condimentum ligula. Morbi quis lacus non mi rutrum finibus id sit amet justo. Fusce rhoncus velit urna, vitae volutpat ex molestie eu. Fusce commodo tincidunt congue. Nunc semper tellus eu turpis fringilla, sit amet laoreet magna suscipit.Etiam eget mi vel lorem lobortis vulputate. Curabitur tortor sapien, tempus non metus eu, eleifend eleifend arcu. Duis leo leo, aliquam sed lorem non, auctor iaculis erat. Mauris ut vehicula dolor, id tincidunt orci. Praesent sed ante sodales, lacinia massa et, venenatis dolor. Mauris placerat velit in nulla mattis, id tincidunt tortor egestas. Donec lacinia posuere lacus ut pulvinar. Fusce ac lacus id tortor fringilla cursus eget id eros. Integer elementum facilisis dolor et tempor."; 
 
@@ -57,3 +58,12 @@ it('testMakeRolesBinary test', () =>{
     expect(makeRolesBinary("Muahaha")).toBeNaN(); 
     expect(makeRolesBinary("Admin")).not.toBeNaN(); 
 })
+
+// kk3609 addition
+// Snapshot of the the Teammember Component
+describe("Teammember Component", () => {
+    it('Matches with the snapshot', () => {
+        const shot = TestRenderer.create(<Teammember/>).toJSON()
+        expect(shot).toMatchSnapshot()
+    })
+});
